+++ src/main/java/is/hi/hbv501g/netkaffi/Services/Implementations/BookingServiceImplementation.java
@@ -33,5 +33,5 @@ public class BookingServiceImplementation implements BookingService {
     public void delete(Booking booking){
         bookingRepository.delete(booking);
     }
-    public Booking findByProductAndStarttime(Product product, int starttime){ return bookingRepository.findByProductAndStarttime(product,starttime); }
+    public Booking findByProductAndStarttime(Product product, long starttime){ return bookingRepository.findByProductAndStarttime(product,starttime); }
 }
